# Add index on numeric_value column for clinic_activity_logs table-created-by-agentic