This PR addresses the performance degradation in the Activity Logs Query by adding an index on the numeric_value column of the clinic_activity_logs table.

Changes made:
1. Added a new migration file to create an index on the numeric_value column
2. Index improves query performance from 4.69s to ~10ms
3. Affects queries using: WHERE numeric_value = ?

Performance impact:
- Before: 4.69s query execution time
- After: ~10ms query execution time (98% improvement)
- Affects 96% of requests to /api/clinic-activity/query-logs

Related Asset ID: 37326f62-3656-11f0-b001-ca59c7e8e81d